The protobard burned with the 10A that ran through the three-phace bridge. It now has been replaced with a soldered breadboard an now the tests are run in it. The problem right now is that the shunt resistor cannot resist more than 10-11A, so that they emit smoke when they are stressed to much.
Now that we have a soldered and fixed version of the bridge the tests for increasing the frequency have been executed. The error encountered was that the fast_cos and fast_sine were programmed to work with degrees but they had as input radians. Due to that the magnetic flux had some sort of poligonal form instead of a circle. When the inputs were changed to angles in degrees the motor work as it should, allowing the code to function up to 32kHz. Just in case, it was set to 30kHz to allow some code manipulation without traslaping signals.